+++ frontend/src/backend/wails/index.ts
@@ -11,9 +11,12 @@ export { CreateFile as createFile } from "./go/main/App";
 export { openFile } from "./openFile";
 export { saveFileAs } from "./saveFileAs";
 
-export { WindowSetTitle as setTitle } from "./runtime/runtime";
+export { WindowSetTitle as setTitle } from "./runtime";
 export { SetIsDirty as setIsDirty } from "./go/main/App";
 export { onExitDirty } from "./trigger";
 export { exitApp } from "./exitApp";
 
 export const testInIframe = true;
+
+export { VersionTag } from "./VersionTag";
+export { InfoDetails } from "./InfoDetails";