+++ frontend/src/components/layout/compact/Header.tsx
@@ -2,6 +2,7 @@ import cn from "classnames";
 import { useStore } from "effector-react";
 import { FC, Fragment } from "react";
 import { ReactComponent as GitHubLogo } from "assets/img/github.svg";
+import { VersionTag } from "backend";
 import { showInfoDialog, ZoomButtons } from "components/common";
 import { EditorTabs, FilesToolbar, FileToolbar } from "components/files";
 import {
@@ -34,6 +35,7 @@ import { ToolButton } from "./ToolButton";
 import { TilesButton } from "./TilesButton";
 import cl from "./Header.module.scss";
 
+const VersionTagC = VersionTag;
 const MUTE = ColorType.MUTE;
 
 interface Props extends ContainerProps {}
@@ -109,14 +111,23 @@ export const Header: FC<Props> = (props) => {
               uiColor={MUTE}
               icon={<GitHubLogo />}
             >
-              GitHub repo <span className={cl.ver}>v{APP_VERSION}</span>
+              GitHub repo
             </TextButton>
             <TextButton
               uiColor={MUTE}
               icon={<svgs.Info />}
               onClick={showInfoDialog}
             >
-              Info
+              Info{" "}
+              <span className={cl.ver}>
+                v{APP_VERSION}
+                {VersionTagC && (
+                  <>
+                    {" "}
+                    (<VersionTagC />)
+                  </>
+                )}
+              </span>
             </TextButton>
           </Toolbar>
         </ButtonDropdown>